Publish_NIMS workflow: Allow updating published releases
When I published 1.1.0, ncipollo/release-action@v1 failed with this error:
```
Error: Error undefined: Tried to update "MeasurementLink Support for Python v1.1.0" which is neither a draft or prerelease. (updateOnlyUnreleased is on)
```

The `Publish_NIMS.yml` workflow doesn't start until you click the
*Publish* button, at which
point the release is no longer a draft.

Signed-off-by: Brad Keryan <brad.keryan@ni.com>