service: Fix compatibility with Python 3.8
In Python 3.8, `weakref.ref` is just a `type` and lower-case `type`
doesn't support subscripting with this version. The fix is to use
`from __future__ import annotations` so that Python 3.8 doesn't try
to parse this type hint.

Also, use `ReferenceType` instead of `ref` in annotations, since it
looks more like a type. (It doesn't really matter because one is an
alias for the other.)